+++ plugins/module_utils/network/saos10/argspec/fps/fps.py
@@ -54,7 +54,7 @@ class FpsArgs(object):  # pylint: disable=R0903
                     "type": "list",
                     "elements": "dict",
                     "options": {
-                        "egress_name": {"type": "str"},
+                        "egress_name": {"type": "str", "required": True},
                         "untagged_dei": {"type": "str", "choices": ["enabled", "disabled"]},
                         "untagged_pcp": {
                             "type": "str",
@@ -68,8 +68,8 @@ class FpsArgs(object):  # pylint: disable=R0903
                             "options": {
                                 "push_vid": {"type": "int", "required": True},
                                 "tag": {"type": "int", "required": True},
-                                "no_op": {"type": "str"},
-                                "pop_type": {"type": "str"},
+                                "no_op": {"type": "list", "elements": "str", "choices": ["null"]},
+                                "pop_type": {"type": "list", "elements": "str", "choices": ["null"]},
                                 "push_dei": {"type": "str", "choices": ["enabled", "disabled"]},
                                 "push_pcp": {
                                     "type": "str",
@@ -111,7 +111,7 @@ class FpsArgs(object):  # pylint: disable=R0903
                         },
                     },
                 },
-                "egress_l3_mapped": {"type": "str"},
+                "egress_l3_mapped": {"type": "list", "elements": "str", "choices": ["null"]},
                 "egress_remark_dscp_value": {"type": "int"},
                 "fd_name": {"type": "str"},
                 "flood_containment_profile": {"type": "str"},
@@ -120,7 +120,7 @@ class FpsArgs(object):  # pylint: disable=R0903
                     "type": "list",
                     "elements": "dict",
                     "options": {
-                        "ingress_name": {"type": "str"},
+                        "ingress_name": {"type": "str", "required": True},
                         "untagged_dei": {"type": "str", "choices": ["enabled", "disabled"]},
                         "untagged_pcp": {
                             "type": "str",
@@ -134,8 +134,8 @@ class FpsArgs(object):  # pylint: disable=R0903
                             "options": {
                                 "push_vid": {"type": "int", "required": True},
                                 "tag": {"type": "int", "required": True},
-                                "no_op": {"type": "str"},
-                                "pop_type": {"type": "str"},
+                                "no_op": {"type": "list", "elements": "str", "choices": ["null"]},
+                                "pop_type": {"type": "list", "elements": "str", "choices": ["null"]},
                                 "push_dei": {"type": "str", "choices": ["enabled", "disabled"]},
                                 "push_pcp": {
                                     "type": "str",
@@ -177,20 +177,20 @@ class FpsArgs(object):  # pylint: disable=R0903
                         },
                     },
                 },
-                "ingress_l3_mapped": {"type": "str"},
+                "ingress_l3_mapped": {"type": "list", "elements": "str", "choices": ["null"]},
                 "ingress_remark_dscp_value": {"type": "int"},
                 "logical_port": {"type": "str"},
                 "mac_learning": {"type": "str", "choices": ["enabled", "disabled"]},
                 "map_policy": {"type": "str", "choices": ["outer-tag", "inner-tag", "mpls-tc", "dscp", "dscp-mpls-tc"]},
                 "meter_profile": {"type": "str"},
-                "mpls_pw": {"type": "str"},
+                "mpls_pw": {"type": "list", "elements": "str", "choices": ["null"]},
                 "mtu_size": {"type": "int"},
                 "normalized_vid": {
                     "type": "list",
                     "elements": "dict",
                     "options": {"tag": {"type": "int", "required": True}, "vlan_id": {"type": "int"}},
                 },
-                "other": {"type": "str"},
+                "other": {"type": "list", "elements": "str", "choices": ["null"]},
                 "pfg_group": {
                     "type": "str",
                     "choices": [
@@ -211,7 +211,7 @@ class FpsArgs(object):  # pylint: disable=R0903
                 "queue_group_instance": {"type": "str"},
                 "stats_collection": {"type": "str", "choices": ["on", "off"]},
                 "svlan": {"type": "int"},
-                "uni": {"type": "str"},
+                "uni": {"type": "list", "elements": "str", "choices": ["null"]},
             },
         },
         "state": {"type": "str", "default": "merged", "choices": ["merged", "deleted"]},